switch to unittest
