+++ tests/test_functions.py
@@ -1,34 +1,9 @@
 import requests, difflib, sys
 import re
-import argparse
 from bs4 import BeautifulSoup
 
-def test_print(string):
-    sys.stdout.write("[synbiohub test] ")
-    sys.stdout.write(string)
-    sys.stdout.write("\n")
+from test_arguments import args, test_print
 
-
-parser = argparse.ArgumentParser()
-
-parser.add_argument("--resetalltests", help="reset all tests for requests by saving responses for future comparisons. Should only be run if working on the test suite implementation itself and all tests have passed.",
-                    action = "store_true")
-
-parser.add_argument("--serveraddress",
-                    help="specify the synbiohub server to test.",
-                    default = "http://localhost:7777/")
-
-args = parser.parse_args()
-
-# format it with a slash if the slash is missing
-if args.serveraddress[-1] != '/':
-    args.serveraddress = args.serveraddress + '/'
-
-number_of_errors = 0
-
-
-if args.resetalltests:
-    test_print("resetting all tests by saving every request for future comparisons. You should not have run this unless you are working on the test suite implementation itself and have verified that all tests passed before the reset.")
 
 
 # make html a little more human readable
@@ -39,6 +14,7 @@ def format_html(htmlstring):
 
 # perform a get request
 def get_request(request):
+    
     response = requests.get(args.serveraddress + request)
     
     response.raise_for_status()
@@ -49,35 +25,39 @@ def get_request(request):
 
 # data is the data field for a request
 def post_request(request, data):
-    response = requests.post(agrs.serveraddress + request, data = data)
+    address = args.serveraddress + request
+
+    response = requests.post(address, data = data)
     response.raise_for_status()
     
     content = format_html(response.text)
     return content
 
-# perform and save a get request in previous results
-def save_get_request(request):
-    content = get_request(request)
-
-    filepath = 'previousresults/getrequest_' + request + '.html'
-    
-    with open(filepath, 'w') as rfile:
-        rfile.write(content)
 
 # creates a file path for a given request and request type
 def request_file_path(request, requesttype):
     return 'previousresults/' + requesttype.replace(" ", "") + "_" + request + ".html"
-        
-# check a request against previous results
-# request should the page that was requested, such as /setup
-# requesttype is the type of request performed, either "get request" or "post request"
+
+
 def compare_request(requestcontent, request, requesttype):
+    """ Checks a request against previous results or saves the result of a request.
+request is the endpoint requested, such as /setup
+requesttype is the type of request performed- either 'get request' or 'post request'"""
     # if the global state is to replace all files, do that
     if args.resetalltests:
         with open(request_file_path(request, requesttype), 'w') as rfile:
             rfile.write(requestcontent)
-    else:
-        
+    elif requesttype[0:3] == "get" and request in args.resetgetrequests:
+        filename = request_file_path(request, requesttype)
+        test_print("resetting get request " + request + " and saving to file " + filename)
+        with open(filename, 'w') as rfile:
+            rfile.write(requestcontent)
+    elif requesttype[0:4] == "post" and request in args.resetpostrequests:
+        filename = request_file_path(request, requesttype)
+        test_print("resetting post request " + request + " and saving to file " + filename)
+        with open(filename, 'w') as rfile:
+            rfile.write(requestcontent)
+    else:      
         filepath = request_file_path(request, requesttype)
         
         olddata = None
@@ -86,7 +66,7 @@ def compare_request(requestcontent, request, requesttype):
                 olddata=oldfile.read()
         except IOError as e:
             raise Exception("\n[synbiohub test] Could not open previous result for the " + \
-                        requesttype + " " + request + ". If the saved result has not yet been created because it is a new page, please use TODO to create the file.") from e
+                            requesttype + " " + request + ". If the saved result has not yet been created because it is a new page, please use --resetgetrequests [requests] or --resetpostrequests [requests] to create the file.") from e
 
         olddata = olddata.splitlines()
         newdata = requestcontent.splitlines()
@@ -95,17 +75,20 @@ def compare_request(requestcontent, request, requesttype):
 
         # temp variable to detect if we need to print the beginning of the error
         firstchangep = True
+
+        # change list holds the strings to print in an error message
+        changelist = [requesttype, " ", request, " did not match previous results. If you are adding changes to SynBioHub that change this page, please check that the page is correct and update the file using the command line argument --resetgetrequests [requests] and --resetpostrequests [requests].\nThe following is a diff of the new files compared to the old.\n"]
         
         for c in changes:
             if firstchangep:
-                test_print(requesttype + " " + request + " did not match previous results. If you are adding changes to SynBioHub that change this page, please check that the page is correct and update the file using TODO.\nThe following is a diff of the new files compared to the old.\n")
-                global number_of_errors
-                number_of_errors += 1
                 firstchangep = False
 
             # print the diff
-            sys.stdout.write(c)
-            sys.stdout.write("\n")
+            changelist.append(c)
+            changelist.append("\n")
+
+        if not firstchangep:
+            raise ValueError(''.join(changelist))
 
     
 # request- string, the name of the page being requested
@@ -121,8 +104,3 @@ def compare_post_request(request, data):
         request = request[1:]
     compare_request(post_request(request, data), request, "post request")
     
-def exit_tests():
-    # finally, if there were any errors, raise an Exception
-    if number_of_errors > 0:
-        raise ValueError("[synbiohub test] " + str(number_of_errors) + " tests failed.")
-
