+++ tests/test_requirements.txt
@@ -1,2 +1,3 @@
 beautifulsoup4
 requests
+unittest
